AP_Torqeedo: add supoort multi-Torqeedos
