+++ libraries/AP_Torqeedo/AP_Torqeedo_config.h
@@ -5,3 +5,11 @@
 #ifndef HAL_TORQEEDO_ENABLED
 #define HAL_TORQEEDO_ENABLED BOARD_FLASH_SIZE > 1024 && !defined(HAL_BUILD_AP_PERIPH)
 #endif
+
+#ifndef AP_TORQEEDO_BACKEND_DEFAULT_ENABLED
+#define AP_TORQEEDO_BACKEND_DEFAULT_ENABLED HAL_TORQEEDO_ENABLED
+#endif
+
+#ifndef AP_TORQEEDO_TQBUS_ENABLED
+#define AP_TORQEEDO_TQBUS_ENABLED AP_TORQEEDO_BACKEND_DEFAULT_ENABLED
+#endif
